Optionally color the Desynthesis delta/level similar to the ExtendedDesynthesisWindow.
Similar to daemitus's comment in #50, I like to GC turnin things that isn't going to offer me any skill ups, but rather than desynth first, I'd like to turn in first to clean up inventory, and continue collecting, to build up a bunch to desynth at the same time (and thus optimize use of potions/food).  This commit adds the same coloring idea found in the ExtendedDesynthesisWindow to the Item Hover tooltip making it easy to see in the GC turnin (as well as anywhere I hover).