[`EmoteLogServerBarEntry`] New Tweak (#671)
* [EmoteLogServerBarEntry] New Tweak

Added a new Tweak that tracks the status of the Emote Log in the server
bar. This will respond to changes via either the Emotes UI panel, or the slash
command /emotelog. Clicking on the entry in the bar will also toggle it.

* Address PR feedback

Moved DtrBar reference to Service.cs
Removed overridden Dispose
Now using class attributes for tweak name and description